+++ builders/NamespaceBuilder.d.ts
@@ -1,0 +1,5 @@
+export default class NamespaceBuilder {
+    private computeNamespace;
+    private walk;
+    load(directory: string): Promise<void>;
+}