+++ loader/NamespaceLoader.d.ts
@@ -1,0 +1,3 @@
+export default class NamespaceLoader {
+    static load(directory: string): Promise<void>;
+}