+++ models/JobModel.js
@@ -1,3 +1,4 @@
+import { QueryBuilder } from "objection";
 import BaseModel from "../bases/BaseModel";
 import EpochTimestamps from "../models/EpochTimestamps";
 export default class JobModel extends EpochTimestamps(BaseModel) {
@@ -5,4 +6,8 @@ export default class JobModel extends EpochTimestamps(BaseModel) {
     static idColumn = "id";
     static updatedColumn = null;
     static deletedColumn = null;
+    static QueryBuilder = QueryBuilder;
+    $beforeUpdate(opt, queryContext) {
+        // do nothing
+    }
 }