+++ src/builders/JobBuilder.ts
@@ -2,6 +2,7 @@ import Luxon from "@bejibun/utils/facades/Luxon";
 import JobModel from "@/models/JobModel";
 
 export default class JobBuilder {
+    protected queue?: string;
     protected now: number;
     protected availableAt: number;
     protected args: Array<any>;
@@ -10,6 +11,12 @@ export default class JobBuilder {
         this.now = Luxon.DateTime.now().toUnixInteger();
         this.availableAt = this.now;
         this.args = [];
+    }
+
+    public setQueue(queue: string): JobBuilder {
+        this.queue = queue;
+
+        return this;
     }
 
     public dispatch(...args: any): JobBuilder {
@@ -26,7 +33,7 @@ export default class JobBuilder {
 
     public async send(): Promise<void> {
         await JobModel.create({
-            queue: Bun.randomUUIDv7(),
+            queue: this.queue,
             payload: JSON.stringify(this.args),
             available_at: this.availableAt,
             created_at: this.now