+++ src/loader/NamespaceLoader.ts
@@ -1,0 +1,7 @@
+import NamespaceBuilder from "@/builders/NamespaceBuilder";
+
+export default class NamespaceLoader {
+    public static async load(directory: string): Promise<void> {
+        return await new NamespaceBuilder().load(directory);
+    }
+}